+++ buildbot.py
@@ -86,6 +86,20 @@ def install(properties):
     run_command(command)
 
 
+def cmake(properties):
+    build_path = 'build'
+    if os.path.exists(build_path):
+        print("Path '{}' already exists - removing".format(build_path))
+        shutil.rmtree(build_path)
+    os.mkdir(build_path)
+
+    old_cwd = os.getcwd()
+    os.chdir(build_path)
+    run_command(['cmake', '../'], env_ext={'VERBOSE': '1'})
+    run_command(['cmake', '--build', '.'], env_ext={'VERBOSE': '1'})
+    os.chdir(old_cwd)
+
+
 def coverage_settings(options):
     options['required_line_coverage'] = 61.7
 
@@ -108,6 +122,8 @@ def main():
         run_tests(properties)
     elif cmd == 'install':
         install(properties)
+    elif cmd == 'cmake':
+        cmake(properties)
     else:
         print("Unknown command: {}".format(cmd))