+++ NEWS.rst
@@ -6,7 +6,7 @@ detailed list of every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Added cmake build file.
 
 5.0.0
 -----